pyproject.toml
[project]
name = "knowledge-gains-ai"
version = "0.1.0"
description = "OpenAI Agentic Workflow System with Web Search, File Search, Input Interpretation, and Handoff Agents"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pathlib>=1.0.1",
    "pydantic>=2.0.0",
    "asyncio-throttle>=1.0.2",
    "aiohttp>=3.9.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]
